created all application security as credential handler and API permissions

+++ src/main/java/com/buccodev/tech_shop/services/CustomerService.java
@@ -11,6 +11,8 @@ import com.buccodev.tech_shop.utils.dtos.customers_dtos.CustomerResponseDto;
 import com.buccodev.tech_shop.utils.mappers.CustomerMapper;
 import jakarta.transaction.Transactional;
 import org.springframework.data.domain.PageRequest;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
@@ -42,20 +44,22 @@ public class CustomerService {
     }
 
     @Transactional
-    public void updateCustomer(Long id, CustomerRequestUpdateDto requestUpdateDto) {
+    public void updateCustomer(Long id, CustomerRequestUpdateDto requestUpdateDto, Authentication authentication) {
+        validateOrderOwnership(id, authentication);
         var customer = customerRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Customer not found"));
-
         customer.setName(requestUpdateDto.name());
         customer.setPhone(requestUpdateDto.phone());
         customerRepository.save(customer);
     }
 
-    public void deleteCustomerById (Long id) {
+    public void deleteCustomerById (Long id, Authentication authentication) {
+        validateOrderOwnership(id, authentication);
         var customer = customerRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Customer not found"));
         customerRepository.deleteById(customer.getId());
     }
 
-    public CustomerResponseDto getCustomerById(Long id) {
+    public CustomerResponseDto getCustomerById(Long id, Authentication authentication) {
+        validateOrderOwnership(id, authentication);
         var customer = customerRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Customer not found"));
         return CustomerMapper.toResponseCustomerDto(customer);
     }
@@ -72,9 +76,22 @@ public class CustomerService {
         return customerRepository.findAll(pageRequest).stream().map(CustomerMapper::toResponseCustomerDto).toList();
     }
 
-    public Customer getByEmail(String email) {
-        return customerRepository.findByEmail(email).orElseThrow(() -> new CredentialInvalidException("Invalid credentials"));
+    public CustomerResponseDto getByEmail(String email, Authentication authentication) {
+        var customer = customerRepository.findByEmail(email)
+                .orElseThrow(() -> new CredentialInvalidException("Customer not found"));
+        validateOrderOwnership(customer.getId(), authentication);
+        return CustomerMapper.toResponseCustomerDto(customer);
     }
 
+    private void validateOrderOwnership(Long userId, Authentication authentication) {
 
+        var user = (UserDetails) authentication.getPrincipal();
+
+        if(user instanceof Customer userCustomer) {
+
+            if (!userCustomer.getId().equals(userId)) {
+                throw new CredentialInvalidException("You don't have permission");
+            }
+        }
+    }
 }
