+++ src/main/java/com/buccodev/tech_shop/services/UserSystemService.java
@@ -1,5 +1,8 @@
 package com.buccodev.tech_shop.services;
 
+import com.buccodev.tech_shop.entities.Customer;
+import com.buccodev.tech_shop.entities.Roles;
+import com.buccodev.tech_shop.exceptions.CredentialInvalidException;
 import com.buccodev.tech_shop.exceptions.ResourceNotFoundException;
 import com.buccodev.tech_shop.repository.UserSystemRepository;
 import com.buccodev.tech_shop.utils.dtos.user_system_dto.UserSystemRequestDto;
@@ -7,6 +10,9 @@ import com.buccodev.tech_shop.utils.dtos.user_system_dto.UserSystemResponseDto;
 import com.buccodev.tech_shop.utils.dtos.user_system_dto.UserSystemUpdateRequestDto;
 import com.buccodev.tech_shop.utils.mappers.UserSystemMapper;
 import org.springframework.data.domain.PageRequest;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -15,34 +21,49 @@ import java.util.List;
 public class UserSystemService {
 
     private final UserSystemRepository userSystemRepository;
+    private final PasswordEncoder passwordEncoder;
 
-    public UserSystemService(UserSystemRepository userSystemRepository) {
+    public UserSystemService(UserSystemRepository userSystemRepository, PasswordEncoder passwordEncoder) {
         this.userSystemRepository = userSystemRepository;
+        this.passwordEncoder = passwordEncoder;
     }
 
     public UserSystemResponseDto createUserSystem(UserSystemRequestDto requestUserSystemDto) {
         var userSystem = UserSystemMapper.toUserSystem(requestUserSystemDto);
+        userSystem.setPassword(passwordEncoder.encode(requestUserSystemDto.password()));
         return UserSystemMapper.toUserSystemResponseDto(userSystemRepository.save(userSystem));
     }
 
-    public void updateUserSystem(Long id, UserSystemUpdateRequestDto requestUserSystemDto) {
+    public UserSystemResponseDto createUserSystemAdmin(UserSystemRequestDto requestUserSystemDto) {
+        var userSystem = UserSystemMapper.toUserSystem(requestUserSystemDto);
+        userSystem.setRole(Roles.ADMIN);
+        userSystem.setPassword(passwordEncoder.encode(requestUserSystemDto.password()));
+        return UserSystemMapper.toUserSystemResponseDto(userSystemRepository.save(userSystem));
+    }
+
+    public void updateUserSystem(Long id, UserSystemUpdateRequestDto requestUserSystemDto, Authentication authentication) {
         var userSystem = userSystemRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("User not found"));
 
+        validateOrderOwnership(userSystem.getId(), authentication);
         userSystem.setUsername(requestUserSystemDto.username());
         userSystem.setRole(requestUserSystemDto.role());
         userSystemRepository.save(userSystem);
     }
 
-    public void deleteUserSystemById(Long id) {
+    public void deleteUserSystemById(Long id, Authentication authentication) {
 
         var userSystem = userSystemRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("User not found"));
+
+        validateOrderOwnership(userSystem.getId(), authentication);
 
         userSystemRepository.deleteById(userSystem.getId());
     }
 
-    public UserSystemResponseDto getUserSystemById(Long id) {
+    public UserSystemResponseDto getUserSystemById(Long id, Authentication authentication) {
 
         var userSystem = userSystemRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("User not found"));
+
+        validateOrderOwnership(userSystem.getId(), authentication);
 
         return UserSystemMapper.toUserSystemResponseDto(userSystem);
     }
@@ -61,14 +82,23 @@ public class UserSystemService {
         return userSystemRepository.findAll(pageRequest).stream().map(UserSystemMapper::toUserSystemResponseDto).toList();
     }
 
-    public UserSystemResponseDto findUserSystemsByName(String name) {
+    public UserSystemResponseDto findUserSystemsByName(String name, Authentication authentication) {
 
         var userSystem = userSystemRepository.findByUsername(name).orElseThrow(() -> new ResourceNotFoundException("User not found"));
-
+        validateOrderOwnership(userSystem.getId(), authentication);
         return UserSystemMapper.toUserSystemResponseDto(userSystem);
     }
 
+    private void validateOrderOwnership(Long userId, Authentication authentication) {
 
+        var user = (UserDetails) authentication.getPrincipal();
 
+        if(user instanceof Customer userCustomer) {
+
+            if (!userCustomer.getId().equals(userId)) {
+                throw new CredentialInvalidException("You don't have permission");
+            }
+        }
+    }
 
 }
